requirements.txt
streamlit>=1.24.0
python-dotenv>=0.19.0
google-generativeai>=0.3.0
pdf2image>=1.16.3
Pillow>=9.0.0
pandas==2.2.0
numpy==1.26.3
poppler-utils==0.1.0